+++ TikokulympicBeta/.gitignore
@@ -1,20 +1,107 @@
-## User settings
+## ビルド成果物
+DerivedData/
+build/
+
+## 個人設定
 xcuserdata/
 
-## Obj-C/Swift specific
+## ワークスペース設定
+*.xcworkspace/xcuserdata/
+*.xcworkspace/xcshareddata/WorkspaceSettings.xcsettings
+
+## スキーム管理
+*.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
+*.xcworkspace/xcshareddata/IDEBuildOperationRecords/*.xcbuildoperation
+
+## プロジェクト設定
+*.pbxuser
+!default.pbxuser
+*.mode1v3
+!default.mode1v3
+*.mode2v3
+!default.mode2v3
+*.perspectivev3
+!default.perspectivev3
+
+## Obj-C/Swift固有のファイル
 *.hmap
 
-## App packaging
+## アプリパッケージング
 *.ipa
 *.dSYM.zip
 *.dSYM
 
-## Playgrounds
+## CocoaPods
+Pods/
+
+## Carthage
+Carthage/Build/
+
+## Swift Package Manager
+.build/
+
+## Fastlane
+fastlane/report.xml
+fastlane/Preview.html
+fastlane/screenshots
+fastlane/test_output
+
+## アーカイブ
+*.xcarchive
+
+## プレイグラウンド
 timeline.xctimeline
 playground.xcworkspace
 
-# Swift Package Manager
-.build/
+## SwiftUIプレビュー
+*.designer.swift
+*.swiftui
 
+## シンボリケーション
+*.bcsymbolmap
+
+## 一時ファイル
+*.tmp
+*.temp
+*.swp
+*~.nib
+
+## アセットカタログ
+*.xcassets
+
+## OS固有のファイル
+.DS_Store
+*.DS_Store
+
+## ロックファイル
+*.lock
+
+## ログファイル
+*.log
+
+## カバレッジ
+*.gcno
+
+## プロファイリング
+*.profdata
+
+## リソース
+*.strings.old
+*.stringsdict.old
+
+## ユニットテストとUIテストの生成ファイル
+test-results.xcresult
+
+## Xcode 11以降
+*.xcresult
+
+## Xcode 13以降
+*.xcattach
+
+## シンボルグラフ
+*.symbols.json
+
+## 設定ファイル
 Config.xcconfig
 GoogleService-Info.plist
+
