+++ TikokulympicBeta/Models/API/RequestProtocol.swift
@@ -10,7 +10,7 @@ import Foundation
 
 protocol ResponseProtocol: Decodable {}
 
-protocol RequestProtocol {
+protocol RequestProtocol : Encodable {
     associatedtype Response: ResponseProtocol
     var baseUrl: String { get }
     var path: String { get }
