+++ TikokulympicBeta/Services/SupabaseService.swift
@@ -1,8 +1,8 @@
 //
-//  SupabaseClientManager.swift
+//  SupabaseService.swift
 //  TikokulympicBeta
 //
-//  Created by 株丹優一郎 on 2024/09/10.
+//  Created by 株丹優一郎 on 2024/09/18.
 //
 
 import Foundation
@@ -22,5 +22,16 @@ class SupabaseClientManager {
 
         client = SupabaseClient(supabaseURL: URL(string: supabaseURL)!, supabaseKey: supabaseApiKey)
         debugPrint("SupabaseClient initialized")
+    }
+    
+    func getAccessToken() async throws -> String {
+        guard let session = try await client?.auth.session else {
+            print("トークン取得失敗: ユーザーがサインインしていません")
+            throw NSError(domain: "SupabaseClientManager", code: 0, userInfo: [NSLocalizedDescriptionKey : "No user session found"])
+        }
+        let accessToken = session.accessToken
+        print("アクセストークン取得成功: \(accessToken)")
+        return accessToken
+        
     }
 }
